Re-organized `backend.rs` into `runtime/{mod.rs,backend.rs,runtime.rs,global_runtime.rs}`.
