+++ analysis/runtime/src/runtime/mod.rs
@@ -1,0 +1,11 @@
+pub mod backend;
+pub mod global_runtime;
+pub mod runtime;
+
+use std::{error::Error, sync::{Mutex, Condvar}};
+
+use once_cell::sync::Lazy;
+
+type AnyError = Box<dyn Error + Send + Sync + 'static>;
+
+static FINISHED: Lazy<(Mutex<bool>, Condvar)> = Lazy::new(|| (Mutex::new(false), Condvar::new()));
